// frontend/src/app/api/model/authTokens.ts
/**
 * Life Management 2.0 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthTokens { 
    accessToken: string;
    refreshToken: string;
    /**
     * Access token lifetime in seconds.
     */
    expiresIn: number;
}